requirements.txt
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
tradingview-screener
prophet
yfinance
matplotlib
python-dotenv
hyperliquid-python-sdk
eth-account
toonify
psycopg2-binary
openai
plotly
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0